ipv orders kan de robot stellages als parameters accepteren

+++ src/main/java/com/example/demo/models/Stellage.java
@@ -15,7 +15,7 @@ public class Stellage extends Object3D implements Updatable {
     private double rotationY = 0;
     private double rotationZ = 0;
 
-    public Stellage(int x, int y, int z) {
+    public Stellage(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
